reactor(StringSplittingLottoGenerator): 이름 변경
- StringSplittingLottoGenerator -> CommaSplittingLottoGenerator
