+++ README.md
@@ -93,7 +93,7 @@
 - [x] Lotto 구현
 - [x] LottoGenerator/RandomLottoGenerator 구현
 - [x] LottoExchanger 구현
-- [ ] StringSplittingLottoGenerator 구현
+- [ ] CommaSplittingLottoGenerator 구현
 - [ ] LottoPrize 구현
 - [ ] WinningResult 구현
 - [ ] LottoAnalyzer 구현
