+++ src/main/java/lotto/lotto/FailureCreatingLottoGeneratorException.java
@@ -8,7 +8,7 @@ public class FailureCreatingLottoGeneratorException extends RuntimeException {
         super(String.format("LottoGenerator 생성 실패했습니다. (입력값: %s, 최소 갯수:%d", lottoNumbers, Lotto.SIZE));
     }
 
-    public FailureCreatingLottoGeneratorException(String value, String delimiter) {
-        super(String.format("LottoGenerator 생성 실패했습니다. (입력값: %s, 최소 갯수: %s", value, delimiter));
+    public FailureCreatingLottoGeneratorException(String value) {
+        super(String.format("LottoGenerator 생성 실패했습니다. (입력값: %s)", value));
     }
 }
